Got the Stage Manger working and now have two sample scene scrolling across the screen, one red and one blue.
Swapping these scenes out for actual textures and actual sprites should be easy enough now---the tricky bit will be placing all the hit boxes for whatever assets we add to the game world--tackle this when we get there, for now it looks like we have a good system to start, that should scale pretty dynamically as we add more stages.

Next to do is to handle the unloading and new loading of dead and new stages.