Have all of the ground work for the red stage in place, just need to work out WTF none of it is working.
Like need to look at the following;

Check on the co-ordinate system, for the background sprites, and make it consistent.

Check on the unload/load new system and come up with a working version.

Make sure that CurrentStage actually gets updated, otherwise we can't track which stage we are in, which is really important for the spacebar activation.

Start with this ^